Move context logic to a custom Provider component
